Add confirm on deletion
사용자가 삭제 버튼을 누르면 confirm 메세지를 출력하고 ok를 눌렀을때에만 서버로 deletion 요청을 보냅니다.

+++ public/js/controller/todoController.js
@@ -59,6 +59,8 @@ export function initTotoController(){
     todoListElement.addEventListener("click",onDeleteBtnClick);
     function onDeleteBtnClick(e){
         if(!e.target.className.includes("delete-btn")) return ;
+        const result = window.confirm("정말 삭제하시겠습니까?");
+        if(!result) return;
         const todoItem=e.target.closest(".todo-item");
         const dbID=todoItem.attributes.dbID.value;
         deleteTodo(handleTodoListChange,dbID);
